% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get}
\alias{cache_get}
\title{Get Value from Cache}
\usage{
cache_get(cache_env, key)
}
\arguments{
\item{cache_env}{Environment, cache storage}

\item{key}{Character string, cache key}
}
\value{
Cached value, or NULL if not found/expired
}
\description{
Get Value from Cache
}
\keyword{internal}
