% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set}
\alias{cache_set}
\title{Set Value in Cache}
\usage{
cache_set(cache_env, key, value)
}
\arguments{
\item{cache_env}{Environment, cache storage}

\item{key}{Character string, cache key}

\item{value}{Any R object to cache}
}
\description{
Set Value in Cache
}
\keyword{internal}
