% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_stats}
\alias{cache_stats}
\title{Get Cache Statistics}
\usage{
cache_stats(cache_env)
}
\arguments{
\item{cache_env}{Environment, cache storage}
}
\value{
List with statistics
}
\description{
Get Cache Statistics
}
\keyword{internal}
