% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{print.bfhllm_cache}
\alias{print.bfhllm_cache}
\title{Print Cache Object}
\usage{
\method{print}{bfhllm_cache}(x, ...)
}
\arguments{
\item{x}{Cache object}

\item{...}{Unused}
}
\description{
Print Cache Object
}
