% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-shiny.R
\name{print.bfhllm_cache_shiny}
\alias{print.bfhllm_cache_shiny}
\title{Print Shiny Cache Object}
\usage{
\method{print}{bfhllm_cache_shiny}(x, ...)
}
\arguments{
\item{x}{Shiny cache object}

\item{...}{Unused}
}
\description{
Print Shiny Cache Object
}
