% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-limiter.R
\name{rate_limiter_cleanup_rpm}
\alias{rate_limiter_cleanup_rpm}
\title{Fjern RPM timestamps ældre end 60 sekunder}
\usage{
rate_limiter_cleanup_rpm()
}
\description{
Fjern RPM timestamps ældre end 60 sekunder
}
\keyword{internal}
