% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-limiter.R
\name{rate_limiter_record_request}
\alias{rate_limiter_record_request}
\title{Registrer et nyt API request}
\usage{
rate_limiter_record_request()
}
\description{
Registrer et nyt API request
}
\keyword{internal}
