% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-limiter.R
\name{rate_limiter_reset_daily_if_needed}
\alias{rate_limiter_reset_daily_if_needed}
\title{Nulstil daglig tæller hvis ny dag (midnat lokal tid)}
\usage{
rate_limiter_reset_daily_if_needed()
}
\description{
Nulstil daglig tæller hvis ny dag (midnat lokal tid)
}
\keyword{internal}
