% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-shiny.R
\name{shiny_cache_clear}
\alias{shiny_cache_clear}
\title{Clear Shiny Session Cache}
\usage{
shiny_cache_clear(session)
}
\arguments{
\item{session}{Shiny session object}
}
\description{
Clear Shiny Session Cache
}
\keyword{internal}
