% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-shiny.R
\name{shiny_cache_get}
\alias{shiny_cache_get}
\title{Get Value from Shiny Session Cache}
\usage{
shiny_cache_get(session, key)
}
\arguments{
\item{session}{Shiny session object}

\item{key}{Character string, cache key}
}
\value{
Cached value, or NULL if not found/expired
}
\description{
Get Value from Shiny Session Cache
}
\keyword{internal}
