% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-shiny.R
\name{shiny_cache_set}
\alias{shiny_cache_set}
\title{Set Value in Shiny Session Cache}
\usage{
shiny_cache_set(session, key, value)
}
\arguments{
\item{session}{Shiny session object}

\item{key}{Character string, cache key}

\item{value}{Any R object to cache}
}
\description{
Set Value in Shiny Session Cache
}
\keyword{internal}
