% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-shiny.R
\name{shiny_cache_stats}
\alias{shiny_cache_stats}
\title{Get Shiny Session Cache Statistics}
\usage{
shiny_cache_stats(session)
}
\arguments{
\item{session}{Shiny session object}
}
\value{
List with statistics
}
\description{
Get Shiny Session Cache Statistics
}
\keyword{internal}
